+++ packages/terra-time-input/src/terra-dev-site/doc/example/TimeInputDefaultTwelveHourSeconds.jsx
@@ -6,7 +6,7 @@ import TimeUtil from 'terra-time-input/lib/TimeUtil';
 class timeInput extends React.Component {
   constructor(props) {
     super(props);
-    this.state = { time: '21:24' };
+    this.state = { time: '21:24:21', hasSeconds: true };
     this.handleTimeChange = this.handleTimeChange.bind(this);
   }
 
@@ -26,6 +26,7 @@ Time Provided:
           value={this.state.time}
           onChange={this.handleTimeChange}
           variant={TimeUtil.FORMAT_12_HOUR}
+          hasSeconds={this.state.hasSeconds}
         />
       </div>
     );