+++ packages/terra-time-input/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+5.0.0 - (June 17, 2019)
+------------------
+### Changed
+* Added 'hasSeconds' prop to allow switching seconds field on and off.
+
 4.1.0 - (June 12, 2019)
 ------------------
 ### Removed
@@ -27,7 +32,7 @@ Unreleased
 3.7.0 - (April 24, 2019)
 ------------------
 ### Changed
-* Replaced string "12-hour" with TimeUtil.FORMAT_12_HOUR in terra dev-site eamples
+* Replaced string "12-hour" with TimeUtil.FORMAT_12_HOUR in terra dev-site samples
 * Moved packages we use solely in doc examples and tests to devDependencies
 
 3.6.0 - (March 26, 2019)